+++ jarvis/api/routers/investment.py
@@ -55,6 +55,8 @@ if not os.getcwd().endswith("Jarvis") or all([models.env.robinhood_user, models.
                                         subject=f"Robinhood Token - {datetime.now().strftime('%c')}",
                                         html_body=rendered)
         if mail_stat.ok:
+            logger.debug(mail_stat.body)
+            logger.info("Token will be reset in 5 minutes.")
             Thread(target=timeout_otp.reset_robinhood, args=(300,)).start()
             raise APIResponse(status_code=HTTPStatus.OK.real,
                               detail="Authentication success. Please enter the OTP sent via email:")